requirement.txt
aiohttp==3.7.4.post0
aiohttp-middlewares==1.1.0
anyio==3.2.1
async-timeout==3.0.1      
asyncclick==7.1.2.3       
asyncio==3.4.3
attrs==21.2.0
bidict==0.21.2
chardet==4.0.0
idna==3.2
importlib-metadata==4.6.0
multidict==5.1.0
python-engineio==4.2.0
python-kasa==0.4.0.dev3
python-socketio==5.3.0
sniffio==1.2.0
typing-extensions==3.10.0.0
websockets==9.1
yarl==1.6.3
zipp==3.5.0